Designing config file.
Designing config files for creating the game. Changing config file and running game should update the game using the config (not done yet). There is also a config for input mapping which should allow to change input binding (not done yet). Figuring out a design which is nice for the config file