finished definition parsing for full ldtk parser
Finished def parsing got to store it next and then work on the level and instance portion of the parser.
Added a way to detect memory leak in the game. There seem to be a bit of memory leak. This will be fixed with while working on the parser.
after full ldtk parser I have to implement separate level ldtk parser.