Correcting physics collider
Correcting physics collision shape and removed implementation where each tile contains their own collider. This result to a really high collider count which is bad for memory. In replace for setting collider in LDTK editor using entity which can easily group collider to a single collider which resolve issue. Rough implementation works, but it is still in progress. Fixed AABB to AABB intersection. Sweep test is currently not producing the correct result. It is also being worked on.

Things that are being worked on.
1) LDTK entities support,
2) LDTK entity collider, and trigger
3) Collider Pivot point support,
4) AABB ABB sweep test correction.