Sprite Batching & Instance Rendering WIP
Improved the rendering and optimized the render code a bit (I will continue optimizing this not done). Sprite batching look like it work and instancing look like it works, but I'm not done quite yet. Got to set up the instance buffer to pass to the vertex buffer and fix a bit of the todos.
I added two new thread UI and audio.
audio has a small implementation of playing audio this will be change for a complete implementation (low level).
a lot of old script need to be removed as well which will be removed step by step