improved game loop.
happy with the rough implementation on the game loop. Might move windows function call to hardware intrinsic function call next, then clean up. After this I will be working on the the rendering and game loop syncing. when using spall profiler it is evident that there are a bit of issue/s.