moved away from window query
moved to rdtsc for timer for better accuracy. going to fixed the issue when there is a new shared sprite batch added and sync with the renderer it take 11.8 ms right now which is will be possible take a lot of time loading texture, shader, etc.. this will go other the game loop 2x then normal which will cause a jitter. this will be noticeable on a lower hz monitor.