small improvement on ecs.
Going to work on View.  s storage for querying. I will work on the querying next after i am done optimizing the solution so far. Currently removed if statement and using bitwise operation . Also reducing function call in the ecs solution. If the ecs call a function it will be inline.  Trying to make ecs function leaf function as much as possible.