safety check for remove
added compile time safety remove component, since removing component of a entity that isn't added will cause a undefined behavior and ruin the order of the group. If user doesn't actually know if the entity has the component that it want to remove then then can pass true on safety.