added entity version incrementing
forgot version incrementing for the entity store. Added version incrementing version get incremented when component is accessed or modified regardless. Updated is alive entity to check the version and dirty version. Version only take up a byte so we added wrap around so when it reaches max it wrap back to zero.