fixed swap and some other major change
grouping and subgrouping now should work, but chose to make the user register then call group or sub group these function should be called first before any manipulation to the sparse sets such as adding component or removing. improved and simplified the component sparse also decided not to go for a generational type id generator to use for look up in the dynamic sparse set array and stick with map, since it quite difficult to create in this language it seems and contains some hidden bug that crash some time (rarely), but that is still bad. I sticked with a map for this case for component info struct. Removed memset on sparse set. Added a few todo for next push. I will also probably work on the system. This will be very minimal on next push and the todo function/s.