Horizontal movement work, but still WIP & Input System
Going to eventually add a way to parse a input file to allow user to remap the input for the game. When i finish the input system and I'm happy with the design  Sensitivity is implemented, but dead and gravity is not done yet on the Input system. The input value snap back to zero when input is release which isn't desirable. Got the player moving with the small bit of the input system  (Horizontal movement)