working progress on the render side
Working progress on render order per instance rather then per batch. The sorting is done in the render thread. It should handle 500 sprite no problem, It create a temp slice using the temp allocator, since the life time is only a cycle so it can be rendered then it is removed. If there is a bottleneck on the render order sorting on the render thread i will implement  a spatial hashing or quadtree/octree. Got color blend to work per instance.