working progress on physics
the physics looks like it is working correctly. the acceleration on the y axis isn't quite right, but it works. Going to fix this later. Also need to work on the TODO on the physics later. Going to refactor. Player movement should use Accumulated Force and not Velocity directly and let the physics loop handle the computing the velocity from the force/s.